pyproject.toml
[project]
name = "fuzzball-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.14",
    "langchain-google-genai>=4.2.1",
    "langchain-text-splitters>=1.1.1",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
]
